+++ packages/files/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.15.3](https://github.com/standardnotes/app/compare/@standardnotes/files@1.15.2...@standardnotes/files@1.15.3) (2023-05-08)
+
+**Note:** Version bump only for package @standardnotes/files
+
 ## [1.15.2](https://github.com/standardnotes/app/compare/@standardnotes/files@1.15.1...@standardnotes/files@1.15.2) (2023-05-04)
 
 **Note:** Version bump only for package @standardnotes/files