+++ packages/services/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.59.5](https://github.com/standardnotes/app/compare/@standardnotes/services@1.59.4...@standardnotes/services@1.59.5) (2023-05-08)
+
+**Note:** Version bump only for package @standardnotes/services
+
 ## [1.59.4](https://github.com/standardnotes/app/compare/@standardnotes/services@1.59.3...@standardnotes/services@1.59.4) (2023-05-04)
 
 **Note:** Version bump only for package @standardnotes/services