+++ packages/snjs/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.168.6](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.168.5...@standardnotes/snjs@2.168.6) (2023-05-08)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
 ## [2.168.5](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.168.4...@standardnotes/snjs@2.168.5) (2023-05-05)
 
 **Note:** Version bump only for package @standardnotes/snjs