+++ packages/web/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.158.3](https://github.com/standardnotes/app/compare/@standardnotes/web@3.158.2...@standardnotes/web@3.158.3) (2023-05-08)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.158.2](https://github.com/standardnotes/app/compare/@standardnotes/web@3.158.1...@standardnotes/web@3.158.2) (2023-05-05)
+
+**Note:** Version bump only for package @standardnotes/web
+
 ## [3.158.1](https://github.com/standardnotes/app/compare/@standardnotes/web@3.158.0...@standardnotes/web@3.158.1) (2023-05-05)
 
 **Note:** Version bump only for package @standardnotes/web